requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
cloudinary==1.28.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-request==1.6.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
